Refactoring function tests
